+++ CMakeLists.txt
@@ -19,6 +19,7 @@ set(python_module_name1 gdposteriormodel)
 pybind11_add_module(${python_module_name1} MODULE 
   src/gaussiandeconvolution/nestedsampler/gdposteriormodel_wrapper.cpp
   src/gaussiandeconvolution/nestedsampler/gdposteriormodel.cpp
+  src/gaussiandeconvolution/shared_functions/probability_distributions.cpp
 )
 install(TARGETS ${python_module_name1} DESTINATION ./nestedsampler)
 
@@ -26,8 +27,6 @@ set(python_module_name2 mcmcsampler)
 pybind11_add_module(${python_module_name2} MODULE 
   src/gaussiandeconvolution/mcmcsampler/mcmcsampler_wrapper.cpp
   src/gaussiandeconvolution/mcmcsampler/mcmcsampler.cpp
-  src/gaussiandeconvolution/mcmcsampler/functions.cpp
-  src/gaussiandeconvolution/mcmcsampler/probability_distributions.cpp
 )
 install(TARGETS ${python_module_name2} DESTINATION ./mcmcsampler)
 
@@ -38,5 +37,6 @@ pybind11_add_module(${python_module_name3} MODULE
   src/gaussiandeconvolution/shared_functions/gd_samplers.cpp
   src/gaussiandeconvolution/shared_functions/gd_scorers.cpp
   src/gaussiandeconvolution/shared_functions/global_random_generator.cpp
+  src/gaussiandeconvolution/shared_functions/probability_distributions.cpp
 )
 install(TARGETS ${python_module_name3} DESTINATION ./shared_functions)
