+++ src/gaussiandeconvolution/shared_functions/mcmc_convergence_statistics.h
@@ -1,0 +1,14 @@
+#ifndef MCMC_CONVERGENCE_STATISTICS
+#define MCMC_CONVERGENCE_STATISTICS
+
+#include <vector>
+
+//Effective number of 
+double rstat(std::vector<double>&, int);
+double rstat(std::vector<std::vector<double>>&);
+
+//Effective number of degrees of freedom
+double effnumber(std::vector<double>&, int);
+double effnumber(std::vector<std::vector<double>>&);
+
+#endif
